[cleanup, utils] Split into submodules
