+++ setup.cfg
@@ -8,6 +8,7 @@ ignore = E402,E501,E731,E741,W503
 max_line_length = 120
 per_file_ignores =
     devscripts/lazy_load_template.py: F401
+    yt_dlp/utils/__init__.py: F401, F403
 
 
 [autoflake]
